게시판 추가

+++ src/controller/FreeBoardController.java
@@ -1,5 +1,7 @@
 package controller;
 
+import java.util.Date;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
@@ -17,60 +19,106 @@ import vo.ReviewPage;
 @Controller
 public class FreeBoardController {
 	private FreeBoardService service;
+
 	@Autowired
 	public void setService(FreeBoardService service) {
 		this.service = service;
 	}
-////////////////////////////////////////////////////////////////////////////
+
+	////////////////////////////////////////////////////////////////////////////
 	@RequestMapping("/freeboard.do")
-	public ModelAndView FreeBoardMain(@RequestParam(defaultValue = "1")int page){
+	public ModelAndView FreeBoardMain(@RequestParam(defaultValue = "1") int page) {
 		FreeBoardPage freeboardPage = service.getFreeBoradPage(page);
 		ModelAndView mv = new ModelAndView();
 		mv.setViewName("freeboard_main");
 		mv.addObject("freeboardPage", freeboardPage);
-		
+
 		return mv;
 	}
-	
-	@RequestMapping("/writeFreeBoardForm.do")
-	public String writeForm() {
-		return "write_freeboard_form";
+
+	@RequestMapping("write.do")
+	public String writeForm(String title, String comment, HttpSession session) {
+		FreeBoard freeBoard = new FreeBoard();
+		// String id =(String) session.getAttribute("id");
+		// String password = (String)session.getAttribute("password");
+		freeBoard.setComment(comment);
+		// freeBoard.setId(id);
+		// freeBoard.setPassword(password);
+		// freeBoard.setDate(new Date());
+		if (service.writeComment(freeBoard, session)) {
+			return "freeboardInsert_success";
+		} else {
+			return "freeboardInsert_false";
+		}
 	}
-	
+
+	@RequestMapping("/modifyForm.do")
+	public ModelAndView updateFreeboardForm(int boardNo) {		
+		FreeBoard original = service.select(boardNo);
+		ModelAndView mv = new ModelAndView();
+		mv.setViewName("modifyFreeBoard");
+		mv.addObject("original", original);
+		return mv;
+	}
+
+	@RequestMapping("/modify.do")
+	public String modifyFreeBoard(int boardNo, String comment, HttpSession session) {
+		FreeBoard original = service.select(boardNo);
+		String id = (String) session.getAttribute("id");
+		if (original.getId().equals(id)) {
+			if(service.modifyFreeBoard(original, session)){
+				return "modifyFreedBoard_success";
+			}else{
+				return "modifyFreeBoard_false";
+			}
+			
+		} else {
+			return "modifyFreeBoard_false";
+		}
+	}
+
 	@RequestMapping("/modifyFreeBoardForm.do")
-	public ModelAndView modifyFreeBoardForm(int id){
-		FreeBoard original = service.regardlessOfReadCount(id);
-		
+	public ModelAndView modifyFreeBoardForm(int id) {
+		FreeBoard original = service.select(id);
+
 		ModelAndView mv = new ModelAndView();
 		mv.addObject("original", original);
 		mv.setViewName("modify_freeboard_form");
-		
+
 		return mv;
 	}
-	
+
 	@RequestMapping("/successModify.do")
-	public String successModify(FreeBoard freeBoard, HttpSession session){
+	public String successModify(FreeBoard freeBoard, HttpSession session) {
 		service.modifyFreeBoard(freeBoard, session);
 		return "modify_complete";
 	}
-	
-	@RequestMapping("/delete.do")
-	public String delete(int boardNo, String password){
-		if(service.deleteFreeBoard(boardNo, password)) {
-			return "delete_freeboard_complete";
-		} else {
+
+	@RequestMapping("/deleteFreeBoardForm.do")
+	public String delete(int boardNo, String password, HttpSession session) {
+		FreeBoard original = service.select(boardNo);
+		String id = (String) session.getAttribute("id");
+
+		if (original.getId().equals(id)) {
+			if (service.deleteFreeBoard(boardNo, password)) {
+				return "delete_freeboard_complete";
+			} else {
+				return "delete_freeboard_fail";
+			}
+		}else
+		{
 			return "delete_freeboard_fail";
 		}
 	}
-	
+
 	@RequestMapping("/readRead.do")
-	public ModelAndView read(int id){
+	public ModelAndView read(int id) {
 		FreeBoard freeBoard = service.readFreeBoard(id);
-		
+
 		ModelAndView mv = new ModelAndView();
 		mv.addObject("freeboard", freeBoard);
 		mv.setViewName("read_freeboard");
-		
+
 		return mv;
 	}
 }
